Serial SubTreeCount

+++ src/test/java/ISTTXTest.java
@@ -227,9 +227,11 @@ public class ISTTXTest {
                 return name.matches( "output_T[0-9]*.txt" );
             }
         } );
-        for ( final File file : files ) {
-            if ( !file.delete() ) {
-                System.err.println( "Can't remove " + file.getAbsolutePath() );
+        if(files != null) {
+            for (final File file : files) {
+                if (!file.delete()) {
+                    System.err.println("Can't remove " + file.getAbsolutePath());
+                }
             }
         }
         final File folder2 = new File("C:\\Users\\DELL\\PycharmProjects\\IST_TDSL_Stats");
@@ -240,9 +242,11 @@ public class ISTTXTest {
                 return name.matches( "stats_T[0-9]*.csv" );
             }
         } );
-        for ( final File file : files2 ) {
-            if ( !file.delete() ) {
-                System.err.println( "Can't remove " + file.getAbsolutePath() );
+        if(files2 != null) {
+            for (final File file : files2) {
+                if (!file.delete()) {
+                    System.err.println("Can't remove " + file.getAbsolutePath());
+                }
             }
         }
     }
@@ -250,123 +254,127 @@ public class ISTTXTest {
 
     @Test
     public void complexMultiThreadTest() throws InterruptedException, FileNotFoundException {
+        while (true) {
 //        System.setOut(new PrintStream(new BufferedOutputStream(new FileOutputStream("output.txt"))));
 //        System.setErr(new PrintStream(new BufferedOutputStream(new FileOutputStream("output.txt"))));
-        // remove output_T* files before test.
-        removeOutputFiles();
-        IST myTree = new IST();
-        int numThreads = 10;
-        Random rand = new Random(1);
-        HashSet<Integer> keySet = new HashSet<>();
-        List<Integer> valueList = new ArrayList<>();
-        int amountOfKeys = 200000;
-        int amountOfKeys2 = 100000;
-        while (keySet.size() != amountOfKeys + amountOfKeys2) {
-            keySet.add(rand.nextInt());
-            valueList.add(rand.nextInt());
-        }
-        List<Integer> totalKeyList = new ArrayList<>(keySet);
-        List<Integer> keyList = totalKeyList.subList(0, amountOfKeys);
+            // remove output_T* files before test.
+            removeOutputFiles();
+            IST myTree = new IST();
+            int numThreads = 10;
+            Random rand = new Random(1);
+            HashSet<Integer> keySet = new HashSet<>();
+            List<Integer> valueList = new ArrayList<>();
+            int amountOfKeys = 10000;
+            int amountOfKeys2 = 5000;
+            while (keySet.size() != amountOfKeys + amountOfKeys2) {
+                keySet.add(rand.nextInt());
+                valueList.add(rand.nextInt());
+            }
+            List<Integer> totalKeyList = new ArrayList<>(keySet);
+            List<Integer> keyList = totalKeyList.subList(0, amountOfKeys);
 
-        // INSERTS
-        System.out.println("Starting Inserts...\n");
-        ArrayList<Thread> threads = new ArrayList<>(numThreads);
-        int index = 0;
-        for (int i = 0; i < numThreads; i++) {
-            threads.add(new Thread(new ISTComplexRun("T" + i, myTree, keyList.subList(index,index+(amountOfKeys/numThreads)),
-                    valueList.subList(index,index+(amountOfKeys/numThreads)), new ArrayList<>(), new ArrayList<>(), "insert",true)));
-            index += amountOfKeys/numThreads;
-        }
-        for (int i = 0; i < numThreads; i++) {
-            threads.get(i).start();
-        }
-        for (int i = 0; i < numThreads; i++) {
-            threads.get(i).join();
-        }
-        try {
+            // INSERTS
+            System.out.println("Starting Inserts...\n");
+            ArrayList<Thread> threads = new ArrayList<>(numThreads);
+            int index = 0;
+            for (int i = 0; i < numThreads; i++) {
+                threads.add(new Thread(new ISTComplexRun("T" + i, myTree, keyList.subList(index, index + (amountOfKeys / numThreads)),
+                        valueList.subList(index, index + (amountOfKeys / numThreads)), new ArrayList<>(), new ArrayList<>(), "insert", true)));
+                index += amountOfKeys / numThreads;
+            }
+            for (int i = 0; i < numThreads; i++) {
+                threads.get(i).start();
+            }
+            for (int i = 0; i < numThreads; i++) {
+                threads.get(i).join();
+            }
+            try {
+                myTree.checkLevels();
+                assert myTree.checkRep();
+            } catch (TXLibExceptions.AbortException e) {
+                System.out.println("Check Rep Failed. Exiting.");
+                //System.exit(1);
+            }
+            myTree.debugCheckRebuild();
+            System.out.println("Finished Inserts\n");
+
+            // LOOKUPS
+            System.out.println("Starting Lookups...\n");
+            threads = new ArrayList<>(numThreads);
+            index = 0;
+            for (int i = 0; i < numThreads; i++) {
+                threads.add(new Thread(new ISTComplexRun("T" + i, myTree, keyList.subList(index, index + (amountOfKeys / numThreads)),
+                        valueList.subList(index, index + (amountOfKeys / numThreads)), new ArrayList<>(), new ArrayList<>(), "lookup", true)));
+                index += amountOfKeys / numThreads;
+            }
+            for (int i = 0; i < numThreads; i++) {
+                threads.get(i).start();
+            }
+            for (int i = 0; i < numThreads; i++) {
+                threads.get(i).join();
+            }
+            myTree.debugCheckRebuild();
             myTree.checkLevels();
+            System.out.println("Finished Lookups\n");
+
+            List<Integer> keyListAdd = totalKeyList.subList(amountOfKeys, amountOfKeys + amountOfKeys2);
+            List<Integer> valueListAdd = valueList.subList(amountOfKeys, amountOfKeys + amountOfKeys2);
+            List<Integer> keyListRemove = totalKeyList.subList(0, amountOfKeys2);
+            List<Integer> valueListRemove = valueList.subList(0, amountOfKeys2);
+
+            // MIXED
+            System.out.println("Starting Mixed...\n");
+            threads = new ArrayList<>(numThreads);
+            index = 0;
+            int mixedAmountOfKeys = amountOfKeys - amountOfKeys2;
+            for (int i = 0; i < numThreads; i++) {
+                threads.add(new Thread(new ISTComplexRun("T" + i, myTree,
+                        keyListAdd.subList(index, index + (mixedAmountOfKeys / numThreads)),
+                        valueListAdd.subList(index, index + (mixedAmountOfKeys / numThreads)),
+                        keyListRemove.subList(index, index + (mixedAmountOfKeys / numThreads)),
+                        valueListRemove.subList(index, index + (mixedAmountOfKeys / numThreads)),
+                        "mixed", true)));
+                index += mixedAmountOfKeys / numThreads;
+            }
+            for (int i = 0; i < numThreads; i++) {
+                threads.get(i).start();
+            }
+            for (int i = 0; i < numThreads; i++) {
+                threads.get(i).join();
+            }
+            myTree.debugCheckRebuild();
+            myTree.checkLevels();
+            System.out.println("Finished Mixed\n");
+
+            keyList = totalKeyList.subList(amountOfKeys2, amountOfKeys + amountOfKeys2);
+            valueList = totalKeyList.subList(amountOfKeys2, amountOfKeys + amountOfKeys2);
+
             assert myTree.checkRep();
-        } catch (TXLibExceptions.AbortException e){
-            System.out.println("Check Rep Failed. Exiting.");
-            //System.exit(1);
-        }
-        myTree.debugCheckRebuild();
-        System.out.println("Finished Inserts\n");
+            
+            // REMOVES
+            System.out.println("Starting Removes...\n");
+            threads = new ArrayList<>(numThreads);
+            index = 0;
+            for (int i = 0; i < numThreads; i++) {
+                threads.add(new Thread(new ISTComplexRun("T" + i, myTree, keyList.subList(index, index + (amountOfKeys / numThreads)),
+                        valueList.subList(index, index + (amountOfKeys / numThreads)), new ArrayList<>(), new ArrayList<>(), "remove", true)));
+                index += amountOfKeys / numThreads;
+            }
+            for (int i = 0; i < numThreads; i++) {
+                threads.get(i).start();
+            }
+            for (int i = 0; i < numThreads; i++) {
+                threads.get(i).join();
+            }
+            myTree.debugCheckRebuild();
+            myTree.checkLevels();
+            System.out.println("Finished Removes\n");
 
-        // LOOKUPS
-        System.out.println("Starting Lookups...\n");
-        threads = new ArrayList<>(numThreads);
-        index = 0;
-        for (int i = 0; i < numThreads; i++) {
-            threads.add(new Thread(new ISTComplexRun("T" + i, myTree, keyList.subList(index,index+(amountOfKeys/numThreads)),
-                    valueList.subList(index,index+(amountOfKeys/numThreads)), new ArrayList<>(), new ArrayList<>(),"lookup",true)));
-            index += amountOfKeys/numThreads;
-        }
-        for (int i = 0; i < numThreads; i++) {
-            threads.get(i).start();
-        }
-        for (int i = 0; i < numThreads; i++) {
-            threads.get(i).join();
-        }
-        myTree.debugCheckRebuild();
-        myTree.checkLevels();
-        System.out.println("Finished Lookups\n");
+            System.out.println("Test is done!\n");
 
-        List<Integer> keyListAdd = totalKeyList.subList(amountOfKeys, amountOfKeys + amountOfKeys2);
-        List<Integer> valueListAdd = valueList.subList(amountOfKeys, amountOfKeys + amountOfKeys2);
-        List<Integer> keyListRemove = totalKeyList.subList(0, amountOfKeys2);
-        List<Integer> valueListRemove = valueList.subList(0, amountOfKeys2);
+            System.out.println("Num Of Aborts = " + TX.abortCount);
 
-        // MIXED
-        System.out.println("Starting Mixed...\n");
-        threads = new ArrayList<>(numThreads);
-        index = 0;
-        int mixedAmountOfKeys = amountOfKeys - amountOfKeys2;
-        for (int i = 0; i < numThreads; i++) {
-            threads.add(new Thread(new ISTComplexRun("T" + i, myTree,
-                    keyListAdd.subList(index,index+(mixedAmountOfKeys/numThreads)),
-                    valueListAdd.subList(index,index+(mixedAmountOfKeys/numThreads)),
-                    keyListRemove.subList(index,index+(mixedAmountOfKeys/numThreads)),
-                    valueListRemove.subList(index,index+(mixedAmountOfKeys/numThreads)),
-                    "mixed",true)));
-            index += mixedAmountOfKeys/numThreads;
         }
-        for (int i = 0; i < numThreads; i++) {
-            threads.get(i).start();
-        }
-        for (int i = 0; i < numThreads; i++) {
-            threads.get(i).join();
-        }
-        myTree.debugCheckRebuild();
-        myTree.checkLevels();
-        System.out.println("Finished Mixed\n");
-
-        keyList = totalKeyList.subList(amountOfKeys2, amountOfKeys + amountOfKeys2);
-        valueList = totalKeyList.subList(amountOfKeys2, amountOfKeys + amountOfKeys2);
-
-        // REMOVES
-        System.out.println("Starting Removes...\n");
-        threads = new ArrayList<>(numThreads);
-        index = 0;
-        for (int i = 0; i < numThreads; i++) {
-            threads.add(new Thread(new ISTComplexRun("T" + i, myTree, keyList.subList(index,index+(amountOfKeys/numThreads)),
-                    valueList.subList(index,index+(amountOfKeys/numThreads)), new ArrayList<>(), new ArrayList<>(),"remove",true)));
-            index += amountOfKeys/numThreads;
-        }
-        for (int i = 0; i < numThreads; i++) {
-            threads.get(i).start();
-        }
-        for (int i = 0; i < numThreads; i++) {
-            threads.get(i).join();
-        }
-        myTree.debugCheckRebuild();
-        myTree.checkLevels();
-        System.out.println("Finished Removes\n");
-
-        System.out.println("Test is done!\n");
-
-        System.out.println("Num Of Aborts = " + TX.abortCount);
-
     }
 
     public class ISTSimpleRun implements Runnable {
